Work on Checks
After discussion, there should be an option.  Can be configured to default to on.

The state on whether checks should be carried out should be in the database.
This means that different ways of actually processing it can be done in the future, and there will be a SSOT of what should be done.

ps. Also fixed a small bug in checks class