Name DB Constraints
(we need names on them as a bit of work I'm about to do needs to change one. And it's generally good practice to have them.
https://alembic.sqlalchemy.org/en/latest/naming.html   )

Normally we should NOT alter existing DB migrations! But in this case it's a problem we need to fix, and there is no data in the database we want to keep, so we are ok with it!